Added Spotify CSV File & Artist Mappings
I moved the Artist entity files into another folder to not confuse the current setup. My mappings are not working as they should, but this shouldn't effect the rest of the project as the localhost:8085/hello is still working as it should. I will continue to work on it later today.